# infrastructure/swagger/openapi.yaml
openapi: "3.0.3"
info:
  title: Fund Transfer System API
  description: |
    Multi-service PHP/Symfony fund transfer platform.
    All endpoints are accessible through NGINX on port 80.
    Each service is routed by path prefix.
  version: "1.0.0"
  contact:
    name: Fund Transfer System
servers:
  - url: http://localhost
    description: Local Docker environment (via NGINX)
...
security:
  - BearerAuth: []
paths:
...
  /user/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              name: Alice Smith
              email: alice@example.com
              password: secret123
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Email already registered.
                code: 409
  /user/login:
    post:
      tags: [Auth]
      summary: Login and receive a JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: alice@example.com
              password: secret123
      responses:
        "200":
          description: Login successful — JWT returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Health endpoints ──────────────────────────────────────────────────────
...
  /user:
    get:
      tags: [Users]
      summary: List all users
      responses:
        "200":
          description: Array of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      tags: [Users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              name: Alice Smith
              email: alice@example.com
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /user/{id}:
    get:
      tags: [Users]
      summary: Get a user by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Accounts ──────────────────────────────────────────────────────────────
...
  /transaction:
    post:
      tags: [Transactions]
      summary: Initiate a fund transfer between two accounts
      description: |
        Creates a pending transaction, debits the source account,
        credits the destination account, and records both ledger entries.
        Returns the completed (or failed) transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
            example:
              sourceAccountId: 1
              destinationAccountId: 2
              amount: "100.00"
              currency: USD
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /transaction/{id}:
    get:
      tags: [Transactions]
      summary: Get a transaction by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Ledger ────────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: integer
      required: [error, code]
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          maxLength: 100
          example: Alice Smith
        email:
          type: string
          format: email
          maxLength: 255
          example: alice@example.com
        password:
          type: string
          minLength: 6
          example: secret123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          example: secret123
    TokenResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    CreateUserRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          maxLength: 100
          example: Alice Smith
        email:
          type: string
          format: email
          maxLength: 255
          example: alice@example.com
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Alice Smith
        email:
          type: string
          example: alice@example.com
        createdAt:
          type: string
          format: date-time
...
    CreateTransactionRequest:
      type: object
      required: [sourceAccountId, destinationAccountId, amount, currency]
      properties:
        sourceAccountId:
          type: integer
          example: 1
        destinationAccountId:
          type: integer
          example: 2
        amount:
          type: string
          example: "100.00"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
    Transaction:
      type: object
      properties:
        id:
          type: integer
          example: 1
        sourceAccountId:
          type: integer
          example: 1
        destinationAccountId:
          type: integer
          example: 2
        amount:
          type: string
          example: "100.00"
        currency:
          type: string
          example: USD
        status:
          type: string
          enum: [pending, completed, failed]
          example: completed
        createdAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Invalid JSON body.
            code: 400
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Invalid or expired token.
            code: 401
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Resource not found.
            code: 404
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: An unexpected error occurred.
            code: 500